pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spotify-llm-controller"
version = "1.0.0"
description = "Spotify LLM Controller for AI-powered music control"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "openai>=1.0.0",
    "fastapi>=0.95.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.20.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
]
dev = [
    "spotify-llm-controller[test]",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/spotify_llm_controller"]